fix: dashboard component
