+++ src/pages/dashboard/common/delete-dialog/delete-dialog.scss
