+++ src/pages/dashboard/common/delete-dialog/delete-dialog.tsx
