--- src/pages/dashboard/common/index.ts
@@ -1,5 +1,0 @@
-import LoadBotPreview from './local';
-
-import './index.scss';
-
-export default LoadBotPreview;
