+++ src/pages/dashboard/common/info-panel.tsx
