--- src/pages/dashboard/common/local-footer.tsx
@@ -1,40 +1,0 @@
-import React from 'react';
-import { observer } from 'mobx-react-lite';
-
-import { Button } from '@deriv-com/ui';
-
-import { useStore } from '@/hooks/useStore';
-import { localize } from '@/utils/tmp/dummy';
-
-import './index.scss';
-
-const LocalFooter = observer(() => {
-    const { ui } = useStore();
-    const { load_modal, dashboard } = useStore();
-    const { is_open_button_loading, loadFileFromLocal, setLoadedLocalFile, toggleLoadModal } = load_modal;
-    const { setPreviewOnPopup } = dashboard;
-
-    const { is_mobile } = ui;
-    const Wrapper = is_mobile ? Button.Group : React.Fragment;
-    return (
-        <Wrapper>
-            {is_mobile && (
-                <Button text={localize('Cancel')} onClick={() => setLoadedLocalFile(null)} has_effect secondary large />
-            )}
-            <Button
-                text={localize('Open')}
-                onClick={() => {
-                    loadFileFromLocal();
-                    setPreviewOnPopup(false);
-                    toggleLoadModal();
-                }}
-                is_loading={is_open_button_loading}
-                has_effect
-                primary
-                large
-            />
-        </Wrapper>
-    );
-});
-
-export default LocalFooter;
