--- src/pages/dashboard/common/recent.tsx
@@ -1,103 +1,0 @@
-import React from 'react';
-import { observer } from 'mobx-react-lite';
-
-import { Text } from '@deriv-com/ui';
-
-import MobileWrapper from '@/components/shared_ui/mobile-wrapper';
-import { getSavedWorkspaces } from '@/external/bot-skeleton';
-import { useStore } from '@/hooks/useStore';
-import { Localize, localize } from '@/utils/tmp/dummy';
-
-import DeleteDialog from './delete-dialog';
-import RecentWorkspace from './recent-workspace';
-import SaveModal from './save-modal';
-
-import './index.scss';
-
-type THeader = {
-    label: string;
-    className: string;
-};
-
-const HEADERS: THeader[] = [
-    {
-        label: localize('Bot name'),
-        className: 'bot-list__header__label',
-    },
-    {
-        label: localize('Last modified'),
-        className: 'bot-list__header__time-stamp',
-    },
-    {
-        label: localize('Status'),
-        className: 'bot-list__header__load-type',
-    },
-];
-
-const RecentComponent = observer(() => {
-    const { load_modal, dashboard } = useStore();
-    const { setDashboardStrategies, dashboard_strategies } = load_modal;
-    const { setStrategySaveType, strategy_save_type } = dashboard;
-    const { ui } = useStore();
-    const { is_mobile } = ui;
-    const get_first_strategy_info = React.useRef(false);
-    const get_instacee = React.useRef(false);
-
-    React.useEffect(() => {
-        setStrategySaveType('');
-        const getStrategies = async () => {
-            const recent_strategies = await getSavedWorkspaces();
-            setDashboardStrategies(recent_strategies);
-            if (!get_instacee.current) {
-                get_instacee.current = true;
-            }
-        };
-        getStrategies();
-        //this dependency is used when we use the save modal popup
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [strategy_save_type]);
-
-    React.useEffect(() => {
-        if (!dashboard_strategies?.length && !get_first_strategy_info.current) {
-            get_first_strategy_info.current = true;
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
-
-    if (!dashboard_strategies?.length) return null;
-    return (
-        <div className='load-strategy__container load-strategy__container--has-footer'>
-            <div className='load-strategy__recent'>
-                <div className='load-strategy__recent__files'>
-                    <div className='load-strategy__title'>
-                        <Text size={is_mobile ? 'xs' : 'sm'} weight='bold'>
-                            <Localize i18n_default_text='Your bots:' />
-                        </Text>
-                    </div>
-                    <div className='bot-list__header'>
-                        {HEADERS.map(({ label, className }) => {
-                            return (
-                                <div className={className} key={label}>
-                                    <Text size={is_mobile ? 'xs' : 'sm'} weight='bold'>
-                                        {label}
-                                    </Text>
-                                </div>
-                            );
-                        })}
-                    </div>
-                    <div className='bot-list__wrapper'>
-                        {dashboard_strategies.map((workspace, index) => {
-                            return <RecentWorkspace key={workspace.id} workspace={workspace} index={index} />;
-                        })}
-                    </div>
-                    <DeleteDialog setStrategies={setDashboardStrategies} />
-                    <MobileWrapper>
-                        <SaveModal />
-                    </MobileWrapper>
-                </div>
-            </div>
-        </div>
-    );
-});
-
-export default RecentComponent;
