+++ src/pages/dashboard/common/stop-bot-dialog/stop-bot-modal-content.tsx
