+++ src/pages/dashboard/common/stop-bot-dialog/stop-bot-modal.tsx
@@ -2,7 +2,7 @@ import { observer } from 'mobx-react-lite';
 
 import { useStore } from '@/hooks/useStore';
 
-import StopBotModalContent from '../stop-bot-modal-content';
+import StopBotModalContent from './stop-bot-modal-content';
 
 const StopBotModal = observer(() => {
     const { run_panel, quick_strategy, summary_card } = useStore();
