+++ src/pages/main/common/run-strategy.tsx
